+++ mayan/apps/document_comments/forms.py
@@ -1,0 +1,18 @@
+from __future__ import absolute_import, unicode_literals
+
+from django import forms
+
+from mayan.apps.common.forms import DetailForm
+
+from .models import Comment
+
+
+class DocumentCommentDetailForm(DetailForm):
+
+    class Meta:
+        fields = ('comment',)
+        extra_fields = (
+            {'field': 'submit_date', 'widget': forms.widgets.DateTimeInput},
+            {'field': 'user'},
+        )
+        model = Comment