+++ mayan/apps/document_comments/icons.py
@@ -7,4 +7,5 @@ icon_comment_add = Icon(
     secondary_symbol='plus'
 )
 icon_comment_delete = Icon(driver_name='fontawesome', symbol='times')
+icon_comment_edit = Icon(driver_name='fontawesome', symbol='pencil-alt')
 icon_comments_for_document = Icon(driver_name='fontawesome', symbol='comment')